feat: add bytecode structures, disassemble
I have decided to skip the custom ArrayList
implementation, I have done so a few times already in C,
and the secondary purpose of this project is to learn Zig,
so it makes sense to use the std lib.

I wanted to use a tagged union for the instructions,
but I could not find a way to do a packed array for that.
So I'm following the C implementation for now.

Chapter: 14.2 - 14.4